Fix the strip number determination
The GEMEtaPartition::strip() function can compute the strip number, which requires local coordinates in the eta partition frame of reference. Previously the local coordinates were provided in the chamber frame of reference - with GEMChamber::toLocal() -, leading to an incorrect derivation of the strip number. This commit ensures the right frame of reference is used - using the local point given by GEMEtaPartition::toLocal().